Refactor menu bar scene and view logic
